+++ 4HC3 Assignment 2 - Code/HC3 A2/Withdraw1.xaml.cs
@@ -21,7 +21,7 @@ namespace HC3_A2
     public partial class Withdraw1 : Page
     {
         string bankNumber, pin;
-        int balance1, balance2, balance3;
+        double balance1, balance2, balance3;
 
         public Withdraw1()
         {
@@ -30,9 +30,9 @@ namespace HC3_A2
             System.IO.StreamReader file = new System.IO.StreamReader("./Resources/userinfo.txt");
             bankNumber = file.ReadLine();
             pin = file.ReadLine();
-            balance1 = Convert.ToInt32(file.ReadLine());
-            balance2 = Convert.ToInt32(file.ReadLine());
-            balance3 = Convert.ToInt32(file.ReadLine());
+            balance1 = Convert.ToDouble(file.ReadLine());
+            balance2 = Convert.ToDouble(file.ReadLine());
+            balance3 = Convert.ToDouble(file.ReadLine());
 
             chequing.Content = String.Format("CHEQUING ACCOUNT - 4645516846 - {0:C2}", balance1);
             saving.Content = String.Format("SAVINGS ACCOUNT - 123654128 - {0:C2}", balance2);
